Feat: add idempotency check for node pool create/delete

+++ dags/tpu_observability/utils/node_pool_util.py
@@ -89,6 +89,21 @@ def create(
       "accelerator_type": node_pool.machine_type,
   })
 
+  check_cmd = (
+      f"gcloud container node-pools list "
+      f"--project={node_pool.project_id} "
+      f"--cluster={node_pool.cluster_name} "
+      f"--location={node_pool.location} "
+      f"--filter='name={node_pool.node_pool_name}' "
+      f"--format='value(name)'"
+  )
+
+  existing_name = subprocess.run_exec(check_cmd).strip()
+
+  if existing_name :
+    logging.info(f"Node pool {node_pool.node_pool_name} already exists. Skipping creation.")
+    return
+
   command = (
       f"gcloud container node-pools create {node_pool.node_pool_name} "
       f"--project={node_pool.project_id} "
@@ -112,6 +127,22 @@ def create(
 @task
 def delete(node_pool: Info) -> None:
   """Deletes the GKE node pool using gcloud command."""
+
+  """Check if the node pool is valid."""
+  check_cmd = (
+      f"gcloud container node-pools list "
+      f"--project={node_pool.project_id} "
+      f"--cluster={node_pool.cluster_name} "
+      f"--location={node_pool.location} "
+      f"--filter='name={node_pool.node_pool_name}' "
+      f"--format='value(name)'"
+  )
+
+  existing_name = subprocess.run_exec(check_cmd).strip()
+
+  if not existing_name :
+    logging.info(f"Node pool {node_pool.node_pool_name} already deleted or does not exist. Skipping.")
+    return
 
   command = (
       f"gcloud container node-pools delete {node_pool.node_pool_name} "
